+++ src/core/utils/http-trace-extractor.ts
