+++ src/core/writers/doc-writer.interface.ts
