+++ src/index.ts
@@ -1,5 +1,5 @@
-export { docRequest } from "./core/doc-builder";
-export { setNRestDocsConfig, getNRestDocsConfig } from "./core/config";
+export { docRequest } from "./core/builders/doc-builder";
+export { setNRestDocsConfig, getNRestDocsConfig } from "./core/config/config";
 
 // errors
 export * from "./errors";
